requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
microvenv==2023.5.post1
mysql-connector-python==9.2.0
packaging==24.2
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.20.2
